QHDEV-627 - Move accessibility tests into sub folder and disable chromatic snapshots
